requirements.txt
fastapi
uvicorn
yfinance
pandas
scikit-learn
pymongo
python-dotenv
bcrypt
